trial for different grid sizes

+++ dataloader/pc_dataset.py
@@ -371,7 +371,7 @@ class Rockbolts_sk(data.Dataset):
         return len(self.im_idx)
 
     def __getitem__(self, index):
-        raw_data = np.fromfile(self.im_idx[index], dtype=np.float16).reshape((-1, 3)) # Changed from (-1,4)
+        raw_data = np.fromfile(self.im_idx[index], dtype=np.float16).astype(np.float32).reshape((-1, 3)) # Changed from (-1,4)
         if self.imageset == 'test':
             annotated_data = np.expand_dims(np.zeros_like(raw_data[:, 0], dtype=int), axis=1)
         else:
